src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/document-register-type-extensions-expected.txt

CONSOLE ERROR: line 28: Uncaught TypeError: document.registerElement is not a function
Testing document.registerElement() type extension behaviours through createElement().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
